Major refactor. Tests and reverse_converter probably broken
