--- cram2ont
@@ -1,5 +1,0 @@
-#!/usr/bin/env python3
-import sys
-import cram2ont
-
-sys.exit(cram2ont.main())
