--- ont2cram
@@ -1,5 +1,0 @@
-#!/usr/bin/env python3
-import sys
-import ont2cram
-
-sys.exit(ont2cram.main())
