+++ ont2cram.egg-info/entry_points.txt
@@ -1,0 +1,3 @@
+[console_scripts]
+ont2cram = ont2cram.__main__:main
+
