+++ ont2cram/__init__.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+
+__version__ = '0.0.2'
+__all__ = ['ont2cram', 'cram2ont', 'common', 'test']
+__license__ = 'Apache License 2.0'
+__author__ = 'EBI'
+__url__ = 'https://github.com/EGA-archive/ont2cram'
+__keywords__ = 'ONT HDF Fast5 CRAM'
+__description__ = 'Oxford Nanopore HDF/Fast5 to CRAM conversion tool'
