+++ ont2cram/test_data/single-read-10-files-ena/N108871_20160805_FNFAD19076_MN16457_sequencing_run_S_cinn050816_95456_ch223_read215_strand.fast5
