--- test
@@ -1,5 +1,0 @@
-#!/usr/bin/env python3
-import sys
-import test
-from test import *
-test.main()
